Move decision show one level out
[#152464564]
