+++ app/javascript/views/ScreeningDecisionShow.jsx
